Refactor order handling
